COMP: silence warnings
svn-url: http://viewvc.slicer.org/viewvc.cgi/Slicer4?view=revision&revision=5340
git-svn-id: http://svn.slicer.org/Slicer4/trunk@5340 3bd1e089-480b-0410-8dfb-8563597acbee

+++ Modules/EMSegment/Registration/vtkBSplineRegistrator.cxx
@@ -136,7 +136,8 @@ public:
         (1.0 - m_SamplingRatio) / (registration->GetNumberOfLevels() - 1.0);
 
       metric->
-        SetNumberOfSpatialSamples(samplingRatio * numVoxels);
+        SetNumberOfSpatialSamples(static_cast<unsigned long>
+                                  (samplingRatio * numVoxels));
 
       std::cerr << "       Image Size: " << 
         registration->GetFixedImagePyramid()->GetOutput(level)->
